[SPARK-58629][CONNECT] Move fallback error test to ErrorUtilsSuite
